refactor: split typed agent context sections
