+++ eslint.config.mjs
@@ -438,6 +438,10 @@ export default [
       'src/agent/output-extraction-failures.js',
       'src/agent/agent-trigger-evaluator.js',
       'src/agent/agent-context-sections.js',
+      'src/agent/agent-context-types.js',
+      'src/agent/agent-context-environment-sections.js',
+      'src/agent/agent-context-prompt-sections.js',
+      'src/agent/agent-context-validation-sections.js',
       'src/agent/context-pack-builder.js',
       'src/agent/context-pack-core.js',
       'src/agent/context-pack-selection.js',
